chore: add test for worker initialization
