+++ src/lib/server.ts
@@ -30,12 +30,12 @@ import { feideClient } from "~/services/auth/clients.js";
 import { AuthService } from "~/services/auth/service.js";
 import { CabinService } from "~/services/cabins/service.js";
 import type { Context } from "~/services/context.js";
-import { EventService } from "~/services/events/service.js";
-import { ListingService } from "~/services/listings/service.js";
+import { EventService } from "~/services/events/index.js";
+import { ListingService } from "~/services/listings/index.js";
 import { MailService } from "~/services/mail/index.js";
-import { OrganizationService } from "~/services/organizations/service.js";
-import { PermissionService } from "~/services/permissions/service.js";
-import { UserService } from "~/services/users/service.js";
+import { OrganizationService } from "~/services/organizations/index.js";
+import { PermissionService } from "~/services/permissions/index.js";
+import { UserService } from "~/services/users/index.js";
 import fastifyMessageQueue from "./fastify/message-queue.js";
 import fastifyPrisma from "./fastify/prisma.js";
 import fastifyService from "./fastify/service.js";
