+++ src/services/events/index.ts
@@ -1,0 +1,1 @@
+export { EventService } from "./service.js";
