+++ src/services/permissions/index.ts
@@ -1,0 +1,1 @@
+export { PermissionService } from "./service.js";
